Unify implementations of SystemFonts as much as possible.
